fix: handle logarithm/exponentiation of luminosity in history file
The history_colums.list does not include the non-logarithmic luminosity as "L", but as "luminosity". Because of this, the usual management of logarithmic and non-logarithmic values does not work, necessitating an explicit case for catching this. While the value "L" is available in the profile_columns.list, it coexists with "luminosity" there, leading to potential problems should anyone use this column title.

This patch fixes this by explicitly checking for the existence of "luminosity" in the log files in the case of a key error that could not be resolved by the other cases specified, or exponentiating "log_L" should "luminosity" be asked for but not found.